+++ CHANGELOG.md
@@ -8,10 +8,13 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [0.7.0.1] - 2025-09-15
 
+### Changed
+
+- Optimized zho_check() to scan only first 1000 bytes of input string.
+
 ### Fixed
 
 - Fixed CLI tool opencc-jieba office pptx (temp file/directory creation error) and epub (Windows file/directory access denied error).
-- Optimized zho_check() to scan only first 1000 bytes of input string.
 
 ---